% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-id.R
\name{get_match_ids}
\alias{get_match_ids}
\title{Scraping Match IDs}
\usage{
get_match_ids(region, summoner_name, api_key)
}
\arguments{
\item{region}{Region where summoners are playing.}

\item{summoner_name}{The summoner's name.}

\item{api_key}{Your API key. It should be of the format \code{"RGAPI-xxxx"}.}
}
\value{
A vector of matches id recently played by \code{summoner_name}.

For more details on these variables, please refer to the Riot's official documentation:
\href{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}{https://developer.riotgames.com/apis#match-v5/GET_getMatchIdsByPUUID}.
}
\description{
Scrapes match ids of a given summoner.
}
\details{
The following are the valid values for the inputs:
\describe{
  \item{\code{region}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
}

To get an API key, it is sufficient to have a Riot account of at least level 5 and visit
\href{https://developer.riotgames.com/}{https://developer.riotgames.com/}. Once logged in,
you can generate a key from your dashboard.\cr
}
\seealso{
\code{\link{get_players_in_tier}}
}
\author{
Riccardo Di Francesco
}
