% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sleep}
\alias{sleep}
\title{Waiting for Next Iteration (Internal Use)}
\usage{
sleep(min = 0, max = 0)
}
\arguments{
\item{min}{Minimum amount of seconds to be waited.}

\item{max}{Maximum amount of seconds to be waited.}
}
\description{
It stops the execution of R for a random number of seconds. Useful for web scarping.
}
