% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier-player.R
\name{get_players_in_tier}
\alias{get_players_in_tier}
\title{Scraping Summoners in a Given Rank and Tier}
\usage{
get_players_in_tier(region, queue, tier, rank, api_key)
}
\arguments{
\item{region}{Region where summoners are playing.}

\item{queue}{Queue type.}

\item{tier}{Tier.}

\item{rank}{Rank.}

\item{api_key}{Your API key. It should be of the format \code{"RGAPI-xxxx"}.}
}
\value{
A data frame with the following variables:
  \item{\code{leagueID}}{The id of the league.}
  \item{\code{queueType}}{The queue type.}
  \item{\code{tier}}{The tier.}
  \item{\code{rank}}{The rank.}
  \item{\code{summoner_id}}{The summoner's id.}
  \item{\code{summoner_name}}{The summoner's name.}
  \item{\code{leaguePoints}}{The league points.}
  \item{\code{wins}}{Number of games won by the summoner.}
  \item{\code{losses}}{Number of games lost by the summoner.}
  \item{\code{veteran}}{Whether the summoner is a veteran.}
  \item{\code{inactive}}{Whether the summoner is inactive.}

For more details on these variables, please refer to the Riot's official documentation:
\href{https://developer.riotgames.com/apis#league-exp-v4/GET_getLeagueEntries}{https://developer.riotgames.com/apis#league-exp-v4/GET_getLeagueEntries}.
}
\description{
Scrapes all the summoners belonging to a given rank and tier.
}
\details{
\code{get_players_in_tier} returns, among others, all the summoners' names in a given tier and rank. Then, one can
use \code{\link{get_match_ids}} and \code{\link{get_match_info}} to scrape all the matches of a given summoner,
ideally looping over all the names here obtained.\cr

The following are the valid values for the inputs:
\describe{
  \item{\code{region}}{\code{"br1"}, \code{"eun1"}, \code{"euw1"}, \code{"jp1"}, \code{"kr"}, \code{"la1"}, \code{"la2"}, \code{"na1"}, \code{"oc1"}, \code{"ru"}, \code{"tr1"}}
  \item{\code{queue}}{\code{"RANKED_SOLO_5x5"}, \code{"RANKED_TFT"}, \code{"RANKED_FLEX_SR"}, \code{"RANKED_FLEX_TT"}}
  \item{\code{tier}}{\code{"CHALLENGER"}, \code{"GRANDMASTER"}, \code{"MASTER"}, \code{"DIAMOND"}, \code{"PLATINUM"}, \code{"GOLD"}, \code{"SILVER"}, \code{"BRONZE"}, \code{"IRON"}}
  \item{\code{rank}}{\code{"I"}, \code{"II"}, \code{"III"}, \code{"IV"}}
}

To get an API key, it is sufficient to have a Riot account of at least level 5 and visit
\href{https://developer.riotgames.com/}{https://developer.riotgames.com/}. Once logged in,
you can generate a key from your dashboard.\cr

Imposing a waiting time between each item scraped is useful to avoid hitting the rate limits imposed by Riot.
With a personal API key, it is not possible to get more than 20 requests every 1 second and 100 requests every 2 minutes.
To avoid clashes, set \code{min} and \code{max} so that the function waits a random amount of seconds before starting
scraping the next item. To use a deterministic amount of second, set \code{min} and \code{max} to the same number.
}
\seealso{
\code{\link{get_match_ids}}, \code{\link{get_match_info}}
}
\author{
Riccardo Di Francesco
}
